Implemented Vector2D

+++ CannonBall/app/src/main/java/coursework/mobile/cannonball/Constants.java
@@ -20,5 +20,5 @@ public interface Constants {
     int    CANNON_WIDTH  = 350;
     double DEFAULT_ANGLE = 0.0F;
     int    BACKGROUND_COLOR = Color.parseColor("#006400");
-    int    BALL_SIZE = 50;
+    float  BALL_SIZE = 50.0F;
 }
